+++ rbx_binary/src/deserializer/state.rs
@@ -1,4 +1,4 @@
-use std::{borrow::Cow, collections::VecDeque, convert::TryInto, io::Read};
+use std::{collections::VecDeque, convert::TryInto, io::Read};
 
 use ahash::{HashMap, HashMapExt, HashSet, HashSetExt};
 use rbx_dom_weak::{
@@ -140,7 +140,7 @@ fn find_canonical_property<'de>(
             }
 
             // TODO: Do we need an additional fix here?
-            let canonical_name = &descriptors.canonical.name;
+            let canonical_name = descriptors.canonical.name;
             let canonical_type = descriptors.canonical.data_type.ty();
             let migration = match &descriptors.canonical.kind {
                 PropertyKind::Canonical {
@@ -154,7 +154,7 @@ fn find_canonical_property<'de>(
             );
 
             Some(CanonicalProperty {
-                name: canonical_name.as_ref().into(),
+                name: canonical_name.into(),
                 ty: canonical_type,
                 migration,
             })
@@ -181,25 +181,26 @@ fn find_canonical_property<'de>(
 
 fn add_property(instance: &mut Instance, canonical_property: &CanonicalProperty, value: Variant) {
     if let Some(PropertySerialization::Migrate(migration)) = canonical_property.migration {
-        let new_property_name = &migration.new_property_name;
         let old_property_name = canonical_property.name;
+        match migration.perform(&value) {
+            Ok(new_value) => {
+                for &new_property_name in migration.new_property_names() {
+                    if !instance.builder.has_property(new_property_name) {
+                        log::trace!(
+                                "Attempting to migrate property {old_property_name} to {new_property_name}"
+                            );
 
-        if !instance.builder.has_property(new_property_name) {
-            log::trace!(
-                "Attempting to migrate property {old_property_name} to {new_property_name}"
-            );
-            match migration.perform(&value) {
-                Ok(new_value) => {
-                    instance.builder.add_property(new_property_name, new_value);
-                    log::trace!(
-                        "Successfully migrated property {old_property_name} to {new_property_name}"
-                    );
+                        instance
+                            .builder
+                            .add_property(new_property_name, new_value.clone());
+                    }
                 }
-                Err(e) => {
-                    log::warn!("Failed to migrate property {old_property_name} to {new_property_name} because: {e}");
-                }
-            };
-        }
+            }
+
+            Err(e) => {
+                log::warn!("Failed to migrate property {old_property_name} because: {e}");
+            }
+        };
     } else {
         instance
             .builder
@@ -402,7 +403,7 @@ impl<'db, R: Read> DeserializerState<'db, R> {
             for instance in instances {
                 let binary_string = chunk.read_binary_string()?;
                 let value = match std::str::from_utf8(binary_string) {
-                    Ok(value) => Cow::Borrowed(value),
+                    Ok(value) => value.to_owned(),
                     Err(_) => {
                         log::warn!(
                             "Performing lossy string conversion on property {}.{} because it did not contain UTF-8.
@@ -411,7 +412,7 @@ This may cause unexpected or broken behavior in your final results if you rely o
                             prop_name
                         );
 
-                        String::from_utf8_lossy(binary_string)
+                        String::from_utf8_lossy(binary_string).into_owned()
                     }
                 };
                 instance.builder.set_name(value);
@@ -439,7 +440,7 @@ This may cause unexpected or broken behavior in your final results if you rely o
                     for instance in instances {
                         let binary_string = chunk.read_binary_string()?;
                         let value = match std::str::from_utf8(binary_string) {
-                            Ok(value) => Cow::Borrowed(value),
+                            Ok(value) => value.to_owned(),
                             Err(_) => {
                                 log::warn!(
                             "Performing lossy string conversion on property {}.{} because it did not contain UTF-8.
@@ -448,11 +449,11 @@ This may cause unexpected or broken behavior in your final results if you rely o
                                     property.name
                                 );
 
-                                String::from_utf8_lossy(binary_string)
+                                String::from_utf8_lossy(binary_string).into_owned()
                             }
                         };
 
-                        add_property(instance, &property, value.as_ref().into());
+                        add_property(instance, &property, value.into());
                     }
                 }
                 VariantType::ContentId => {
@@ -1234,6 +1235,35 @@ rbx-dom may require changes to fully support this property. Please open an issue
                         add_property(instance, &property, net_asset.into());
                     }
                 }
+                // On 2026-06-30, Roblox began storing Tags in the SharedString index.
+                VariantType::Tags => {
+                    let values = chunk.read_interleaved_u32_array(instances.len())?;
+
+                    for (value, instance) in values.zip(instances) {
+                        let tags_shared_string = self
+                            .shared_strings
+                            .get(value as usize)
+                            .ok_or_else(|| InnerError::InvalidPropData {
+                                type_name: type_name.to_string(),
+                                prop_name: prop_name.to_owned(),
+                                valid_value: "a valid Tags SharedString index",
+                                actual_value: format!("{value:?}"),
+                            })?;
+
+                        add_property(
+                            instance,
+                            &property,
+                            Tags::decode(tags_shared_string.data())
+                                .map_err(|_| InnerError::InvalidPropData {
+                                    type_name: type_name.to_string(),
+                                    prop_name: prop_name.to_owned(),
+                                    valid_value: "a list of valid null-delimited UTF-8 strings",
+                                    actual_value: "invalid UTF-8".to_string(),
+                                })?
+                                .into(),
+                        );
+                    }
+                }
                 invalid_type => {
                     return Err(InnerError::PropTypeMismatch {
                         type_name: type_name.to_string(),
@@ -1306,13 +1336,15 @@ rbx-dom may require changes to fully support this property. Please open an issue
                         });
                     }
 
+                    let optional_markers = chunk.read_slice(instances.len())?;
                     let values = x
                         .zip(y)
                         .zip(z)
                         .map(|((x, y), z)| Vector3::new(x, y, z))
                         .zip(rotations)
-                        .map(|(position, rotation)| {
-                            if chunk.read_u8().ok()? == 0 {
+                        .zip(optional_markers)
+                        .map(|((position, rotation), &marker)| {
+                            if marker == 0 {
                                 None
                             } else {
                                 Some(CFrame::new(position, rotation))