+++ rbx_types/src/referent.rs
@@ -196,8 +196,8 @@ mod serde_test {
     fn non_human() {
         let value = Ref::new();
 
-        let ser = bincode::serialize(&value).unwrap();
-        let de = bincode::deserialize(&ser).unwrap();
+        let ser = rmp_serde::to_vec(&value).unwrap();
+        let de = rmp_serde::from_slice(&ser).unwrap();
 
         assert_eq!(value, de);
     }